% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transGFM.R
\name{relative_error}
\alias{relative_error}
\title{Calculate relative error between estimated and true matrices}
\usage{
relative_error(M_hat, M_true)
}
\arguments{
\item{M_hat}{Estimated matrix}

\item{M_true}{True matrix}
}
\value{
Relative Frobenius norm error
}
\description{
Calculate relative error between estimated and true matrices
}
\examples{
M_true <- matrix(1:9, 3, 3)
M_hat <- M_true + matrix(rnorm(9, 0, 0.1), 3, 3)
relative_error(M_hat, M_true)
}
